requirements.txt
asgiref==3.7.2
astroid==3.0.3
asttokens==2.4.1
autopep8==2.0.4
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
decorator==5.1.1
dill==0.3.7
Django==4.2.11
django-cors-headers==4.3.1
django-crontab==0.7.1
django-environ==0.11.2
django-filter==23.5
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-solo==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
ghasedakpack==0.1.13
idna==3.6
importlib-metadata==7.0.0
inflection==0.5.1
ipython==8.12.3
isort==5.13.2
itypes==1.2.0
jalali_core==1.0.0
jdatetime==5.0.0
jedi==0.19.1
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
pep8==1.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.12.3
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
zipp==3.17.0
